@@ -9,42 +9,40 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^8.0",
-        "bensampo/laravel-enum": "^3.0",
+        "php": "^8.0.2",
+        "bensampo/laravel-enum": "^5.1",
         "fakerphp/faker": "^1.9.1",
-        "fideloper/proxy": "^4.2",
-        "fruitcake/laravel-cors": "v2.0.4",
         "graylog2/gelf-php": "^1.5",
-        "guzzlehttp/guzzle": "^7.0.1",
-        "laravel/framework": "^v8.83",
-        "laravel/slack-notification-channel": "^2.3",
-        "laravel/telescope": "^4.3",
-        "laravel/tinker": "^2.0",
+        "guzzlehttp/guzzle": "^7.2",
+        "laravel/framework": "^9.2",
+        "laravel/slack-notification-channel": "^2.4",
+        "laravel/telescope": "^4.7",
+        "laravel/tinker": "^2.7",
         "laravel/ui": "^3.0",
         "laravelcollective/html": "^6.1",
-        "league/flysystem-aws-s3-v3": "^1.0",
+        "league/flysystem-aws-s3-v3": "^3.0",
         "maatwebsite/excel": "^3.1",
         "mockery/mockery": "^1.3.1",
         "mpdf/mpdf": "^8.0",
-        "owen-it/laravel-auditing": "^10.0",
-        "phpunit/phpunit": "^9.3",
+        "owen-it/laravel-auditing": "^12.0",
+        "phpunit/phpunit": "^9.5.10",
         "predis/predis": "^1.0",
-        "rappasoft/laravel-authentication-log": "^1.2",
+        "rappasoft/laravel-authentication-log": "^2.0",
         "silber/bouncer": "^v1.0.0-rc.9",
         "spatie/cpu-load-health-check": "^1.0",
         "spatie/icalendar-generator": "2.3.0",
         "spatie/laravel-health": "^1.7",
         "spatie/temporary-directory": "^2.0",
         "stuk/jszip": "^3.0",
-        "symfony/http-kernel": "^5.1.5",
+        "symfony/http-kernel": "^6.0",
         "torann/geoip": "^3.0",
         "yajra/laravel-datatables-oracle": "^9.10"
     },
     "require-dev": {
         "barryvdh/laravel-ide-helper": "^2.12",
-        "facade/ignition": "^2.3.6",
+        "spatie/laravel-ignition": "^1.0",
         "laravel/dusk": "^6.9",
-        "nunomaduro/collision": "^5.0",
+        "nunomaduro/collision": "^6.1",
         "nunomaduro/larastan": "^1.0",
         "roquie/laravel-dusk-select2": "^2.1"
     },
