@@ -14,7 +14,7 @@ use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 
 use Datatables;
-use Carbon\Carbon;
+use Illuminate\Support\Carbon;
 use Spatie\TemporaryDirectory\TemporaryDirectory;
 
 
