@@ -8,7 +8,7 @@ use App\Models\Game;
 use App\Models\Team;
 use App\Models\Gym;
 use App\Models\Schedule;
-use Carbon\Carbon;
+use Illuminate\Support\Carbon;
 
 use Tests\TestCase;
 use Tests\Support\Authentication;
